Change float variables to double for precision

+++ code/2-4.c
@@ -8,8 +8,8 @@ int main()
     printf("b = %.100f\n", b);
     printf("a == b is %d\n", a == b);
 
-    float c = 0.1 + 0.2;
-    float d = 0.3;
+    double c = 0.1 + 0.2;
+    double d = 0.3;
     printf("c = %.100f\n", c);
     printf("d = %.100f\n", d);
     printf("c == d is %d\n", c == d);
